+++ src/resources/lang/en/users.php
@@ -7,5 +7,6 @@ return [
     'verify' => 'Verify',
     'verified' => 'Users registration successfully verified',
     'already_verified' => 'Users registration already verified',
-
+    'profile_image' => 'Profile image',
+    'last_device' => 'Last device',
 ];