+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.0] - 2020-05-29
+### Added
+- profile image handling
+
+### Changed
+- Moved methods from BaseCrudController to traits
+
 ## [0.5.0] - 2020-05-29
 ### Added
 - UsersCrud